Timeouts on chunked transfers were not handled properly
This commit adds a connection abort in case of a timeout happens
during a chunked transfer:
A reasonable tear down and/or sending any content to the client is
impossible as we have already started to transmit data.
Thus it is impossible to predict how a client would treat any 'appened'
error message to the stream anyway - and trailing headers are not
supported by any reasonable HTTP client library out there.

Therefore the connection is just aborted to get back to a "definied
state".